+++ LanCloud/LanCloud.FtpServer/Domain/IO/IFileBit.cs
@@ -3,7 +3,7 @@
     public interface IFileBit
     {
         string Extention { get; }
-        int[] Indexes { get; }
+        byte[] Indexes { get; }
         bool IsTemp { get; }
         string Hash { get; }
         long Length { get; }
