+++ LanCloud/LanCloud.FtpServer/Services/IndexesService.cs
@@ -4,7 +4,7 @@ namespace LanCloud.Services
 {
     public static class IndexesService
     {
-        public static bool Matches(this int[] thisOne, int[] compareTo)
+        public static bool Matches(this byte[] thisOne, byte[] compareTo)
         {
             return thisOne.Length == compareTo.Length &&
                 thisOne.All(a => compareTo.Any(b => a == b));
